+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Fix measure_execution_time decorator for sync methods
 - Add async support to callback_system_event
+- Replace HH_EVENT constants by HmSystemEvent enum
 
 # Version 2023.8.12 (2023-08-27)